+++ packages/form-client/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.5](https://github.com/crossmodel/crossmodel-core/compare/v0.10.4...v0.10.5) (2025-11-02)
+
+**Note:** Version bump only for package @crossmodel/form-client
+
+
+
+
+
+## [0.10.4](https://github.com/crossmodel/crossmodel-core/compare/v0.10.3...v0.10.4) (2025-11-02)
+
+**Note:** Version bump only for package @crossmodel/form-client
+
+
+
+
+
 ## [0.10.3](https://github.com/crossmodel/crossmodel-core/compare/v0.10.2...v0.10.3) (2025-10-31)
 
 **Note:** Version bump only for package @crossmodel/form-client